add error case for checkboxes and move input logic to helper file
